renames ploomber.inline -> ploomber.micro
