+++ tests/test_micro.py
@@ -7,8 +7,9 @@ import matplotlib.pyplot as plt
 import nbformat
 import pytest
 
-from ploomber.exceptions import DAGBuildError
-from ploomber import inline
+from ploomber.exceptions import DAGBuildError, DAGRenderError
+from ploomber import micro
+from ploomber.micro import _capture
 
 
 def ones(input_data):
@@ -27,18 +28,18 @@ def multiply(first, second):
     return first * second
 
 
-@inline.grid(a=[1, 2], b=[3, 4])
+@micro.grid(a=[1, 2], b=[3, 4])
 def add(ones, a, b):
     return ones + a + b
 
 
-@inline.grid(a=[1, 2], b=[3, 4])
-@inline.grid(a=[5, 6], b=[7, 8])
+@micro.grid(a=[1, 2], b=[3, 4])
+@micro.grid(a=[5, 6], b=[7, 8])
 def add_many(ones, a, b):
     return ones + a + b
 
 
-@inline.capture
+@micro.capture
 def plot_ones(ones):
     # tag=plot
     plt.plot(ones)
@@ -48,7 +49,7 @@ def plot_ones(ones):
 
 @pytest.mark.parametrize('parallel', [True, False])
 def test_inline(tmp_directory, parallel):
-    dag = inline.dag_from_functions(
+    dag = micro.dag_from_functions(
         [ones, twos, both],
         params={"ones": {
             "input_data": [1] * 3
@@ -69,7 +70,7 @@ def test_inline(tmp_directory, parallel):
 
 
 def test_inline_with_manual_dependencies(tmp_directory):
-    dag = inline.dag_from_functions(
+    dag = micro.dag_from_functions(
         [ones, twos, multiply],
         output="cache",
         params={"ones": {
@@ -91,11 +92,11 @@ def test_inline_with_manual_dependencies(tmp_directory):
 
 
 def test_inline_grid(tmp_directory):
-    dag = inline.dag_from_functions([ones, add],
-                                    params={"ones": {
-                                        "input_data": [1] * 3
-                                    }},
-                                    output='cache')
+    dag = micro.dag_from_functions([ones, add],
+                                   params={"ones": {
+                                       "input_data": [1] * 3
+                                   }},
+                                   output='cache')
 
     dag.build()
 
@@ -113,11 +114,11 @@ def test_inline_grid(tmp_directory):
 
 
 def test_inline_grid_multiple(tmp_directory):
-    dag = inline.dag_from_functions([ones, add_many],
-                                    params={"ones": {
-                                        "input_data": [1] * 3
-                                    }},
-                                    output='cache')
+    dag = micro.dag_from_functions([ones, add_many],
+                                   params={"ones": {
+                                       "input_data": [1] * 3
+                                   }},
+                                   output='cache')
 
     dag.build()
 
@@ -165,8 +166,9 @@ def test_inline_grid_multiple(tmp_directory):
 
 
 @pytest.mark.parametrize('parallel', [True, False])
-def test_capture(tmp_directory, parallel):
-    dag = inline.dag_from_functions(
+@pytest.mark.parametrize('debug', [None, 'now', 'later'])
+def test_capture(tmp_directory, parallel, debug):
+    dag = micro.dag_from_functions(
         [ones, plot_ones],
         params={"ones": {
             "input_data": [1] * 3
@@ -175,7 +177,7 @@ def test_capture(tmp_directory, parallel):
         parallel=parallel,
     )
 
-    dag.build()
+    dag.build(debug=debug)
 
     ones_ = pickle.loads(Path('cache', 'ones').read_bytes()).to_dict()
     plot_ones_ = pickle.loads(Path('cache', 'plot_ones').read_bytes())
@@ -188,37 +190,38 @@ def test_capture(tmp_directory, parallel):
     assert nb.cells[0].metadata.tags[0] == 'plot'
 
 
-# this fails since it tries to unpickle the THML
-def test_capture_can_return_nothing():
+# this fails since it tries to unpickle the HTML
+@pytest.mark.xfail
+def test_capture_can_return_nothing(tmp_directory):
 
-    @inline.capture
+    @micro.capture
     def first():
         x = 1
         return x
 
     # end nodes don't have to return anything
-    @inline.capture
+    @micro.capture
     def second(first):
         pass
 
-    dag = inline.dag_from_functions([first, second])
+    dag = micro.dag_from_functions([first, second])
     dag.build()
 
 
 def test_capture_debug_now(tmp_directory, monkeypatch):
 
-    @inline.capture
+    @micro.capture
     def number():
         x, y = 1, 0
         x / y
 
-    dag = inline.dag_from_functions([number])
+    dag = micro.dag_from_functions([number])
 
     class MyException(Exception):
         pass
 
     mock = Mock(side_effect=MyException)
-    monkeypatch.setattr(inline, 'debug_if_exception', mock)
+    monkeypatch.setattr(_capture, 'debug_if_exception', mock)
 
     with pytest.raises(MyException):
         dag.build(debug='now')
@@ -234,12 +237,12 @@ def test_capture_debug_now(tmp_directory, monkeypatch):
 
 def test_capture_debug_later(tmp_directory, monkeypatch):
 
-    @inline.capture
+    @micro.capture
     def number():
         x, y = 1, 0
         x / y
 
-    dag = inline.dag_from_functions([number])
+    dag = micro.dag_from_functions([number])
 
     with pytest.raises(DAGBuildError):
         dag.build(debug='later')
@@ -247,20 +250,31 @@ def test_capture_debug_later(tmp_directory, monkeypatch):
     assert Path('number.dump').is_file()
 
 
-# this fails since it tries to unpickle the HTML
-def test_capture_that_depends_on_capture():
+def test_capture_that_depends_on_capture(tmp_directory):
 
-    @inline.capture
+    @micro.capture
     def first():
         x = 1
         return x
 
-    @inline.capture
+    @micro.capture
     def second(first):
-        return first + 1
+        second = first + 1
+        return second
 
-    dag = inline.dag_from_functions([first, second])
+    dag = micro.dag_from_functions([first, second])
     dag.build()
+
+
+@pytest.mark.xfail
+def test_error_when_returning_non_variable_and_capture(tmp_directory):
+
+    @micro.capture
+    def first():
+        return 1
+
+    dag = micro.dag_from_functions([first])
+    dag.build(debug='now')
 
 
 def test_root_node_with_no_arguments(tmp_directory):
@@ -271,7 +285,7 @@ def test_root_node_with_no_arguments(tmp_directory):
     def add(root):
         return root + 1
 
-    dag = inline.dag_from_functions([root, add])
+    dag = micro.dag_from_functions([root, add])
     dag.build()
 
     root_ = pickle.loads(Path('output', 'root').read_bytes())
@@ -282,26 +296,30 @@ def test_root_node_with_no_arguments(tmp_directory):
 
 
 # TODO: also test with grid
-def test_decorated_root_with_input_data():
+def test_decorated_root_with_input_data(tmp_directory):
 
     # this is failing because it's not passing input_data - we're not
     # validating the signature,and just modifying the user's namespace,
     # but we have to, otherwise the error is confusing: "input_data" is not
     # defined
-    @inline.capture
+    @micro.capture
     def root(input_data):
         x = input_data + 1
         return x
 
-    dag = inline.dag_from_functions([root])
-    dag.build()
+    dag = micro.dag_from_functions([root])
+
+    with pytest.raises(DAGRenderError) as excinfo:
+        dag.build()
+
+    assert "'root' function missing 'input_data'" in str(excinfo.value)
 
 
 # TODO: also try with grid
 # NOTE: this is failing because it's trying to unpickle the html
 def test_decorated_root_without_arguments(tmp_directory):
 
-    @inline.capture
+    @micro.capture
     def root():
         x = 1
         return x
@@ -309,7 +327,7 @@ def test_decorated_root_without_arguments(tmp_directory):
     def add(root):
         return root + 1
 
-    dag = inline.dag_from_functions([root, add])
+    dag = micro.dag_from_functions([root, add])
     dag.build()
 
     root_ = pickle.loads(Path('output', 'root').read_bytes())
@@ -324,8 +342,8 @@ def get():
     return df
 
 
-@inline.capture
-@inline.grid(model=[
+@micro.capture
+@micro.grid(model=[
     'RandomForestClassifier',
     'AdaBoostClassifier',
     'ExtraTreesClassifier',
@@ -341,7 +359,7 @@ def fit(get, model):
 
 
 def test_decorated_with_capture_and_grid(tmp_directory):
-    dag = inline.dag_from_functions([get, fit])
+    dag = micro.dag_from_functions([get, fit])
     dag.build()
 
     nb = nbformat.reads(Path('output', 'fit-0.ipynb').read_text(),
@@ -362,15 +380,34 @@ def fn():
 
 
 def test_get_body_statements():
-    assert len(inline.get_body_statements(fn)) == 2
+    assert len(_capture._get_body_statements(fn)) == 2
 
 
-def test_deindents_statements():
-    # we're currently passing statements to IPython with the indentation, this
-    # breaks when using parso.parse to
-    # extract the return statement - the current: cell.source.strip()
-    # only works on simple cases
-    raise NotImplementedError
+def simple():
+    x = 1
+    y = 2
+    return x, y
+
+
+def complex():
+    for i in range(10):
+        for j in range(10):
+            if i + j == 2:
+                print('hello')
+
+
+@pytest.mark.parametrize('fn, expected', [
+    [simple, ['x = 1', 'y = 2', 'return x, y']],
+    [
+        complex,
+        [
+            "for i in range(10):\n    for j in range(10):\n"
+            "        if i + j == 2:\n            print('hello')"
+        ]
+    ],
+])
+def test_deindents_statements(fn, expected):
+    assert _capture._get_body_statements(fn) == expected
 
 
 @pytest.mark.parametrize('source, expected', [
@@ -382,8 +419,4 @@ def test_deindents_statements():
     ['\n    # tag=plot\n    plot.confusion_matrix(y_test, y_pred)\n', 'plot'],
 ])
 def test_parse_tag(source, expected):
-    assert inline.parse_tag(source) == expected
-
-
-def test_aborted_when_task_fails():
-    raise NotImplementedError
+    assert _capture._parse_tag(source) == expected
